src/ccf/requirements_ml.txt
hydra-core
pandas
# SQLAlchemy
torch
pytorch-forecasting